combine simple shouldRender functionality
